Added Settings preference page and plugin.xml
This class represents the settings' preference page we use for this
plugin. This page has two
fields : First off we have a directory dialog determining the cygwin's
path for windows. Then we have a field determining
where the user wants to output the results of his/her used commands. It
could be the screen or a
file. All the values are saved in the Preference Store.

The second field will be removed in the near future cause we want to
determine the output for each command separately.

Finally added plugin.xml adding the above preference page to the
workbench.