Custom Command Creation Wizard : Part I
Added wizard's main page and its controller.

In this page the user can set the newly added
command's nickname, the actual command, the shell's starting directory.
An auto complete
feature is implemented to help the user browse through the prototype
commands quickly. When a
prototype command is finally selected , its description appears
automatically in a text box below
the command to help the new users. Finally, if the user selected to pipe
in a previous wizard,
then in the top of the page we display the current state of the command.
Restrictions : the
command name cannot be duplicate or empty, because we use it in our
inner engine as a search key.