Custom Command Creation Wizard : Part II
Added arguments' page and its controller.

In this page the user can choose one or more
from the arguments available for the command he chose. To make the
process easier for the
experienced users we provide a text field where the user can type the
desired arguments.
Otherwise the user can check multiple check buttons , each one having an
argument, and the
argument's description as a tool tip text.