Custom Command Creation Wizard : Part V
Final Part. Added the wizard's creation class.

This class represents the wizard used to guide the user through the
custom command creation
process . It has 3 pages : a Command Page where the user can specify the
newly added command's
name, the command itself, and the shell's starting directory for the
command. The user also gets
the command's description and an auto complete feature to help him/her
choose from the command
prototypes. The second page is an Argument page, where the user can
select from the available
arguments for the command, and finally, the third page is a Resource
page, where the user can
attach an input and output file to the command or add a pipe. If the
user chooses to pipe the wizard starts
all over again, showing in the first page the state of the command so
far. NOTE : When piping ,
the shell's directory path and the command's nickname, are extracted
from the last part of the
pipe i.e from the last wizard that pops up.