Fixed an issue when piping in the addition wizard.
When piping the user had to input the command's nickname and the
command's shell starting dir in every popping dialog.
Now, whenever he/she inputs a nickname or a shell dir it is transfered
to the next wizard.
The fields in the final wizard are taken into consideration so that the
user can change his/her mind throughout the process.