Refactoring changes
Change the way the controllers and the models are instantiated. A new
singleton class PluginContext was created in order to deal with that.
The method init is called inside the Activator of the plugin to create
all the necessary objects. Getters are provided for other classes who
need the controllers. One of the main differences is that that the
model is now passed as parameter to the constructor of a controller.
As a result the controllers can now be tested more easily.