Fixed a bug in the wizard's resource page.
Explanation of the bug :
 When adding resources to the selected command, the user could just
leave the path's textfield blank and then press ok, resulting in an
empty entry in the resources' table.

Fix :
    Added a modifyListener to the text checking the text's state.
If it is empty it sets an error message which disposes when it is empty
no more. The error message acts as a flag for the process' completion.

  We explicitly call the modify listener of the text on the start. Why
would we like to do that?
The first time the user opens the dialog the path will be always empty.
but since we associated the resource editing along with this class,
we cannot just setErrorMessage(something) in the control creation
since if the user wants to edit an inserted resource file and along the
process changes his mind and just presses ok without changing the path
the dialog won't let him save and continue. Therefore we explicitly
invoke the modify listener and it will display the message in an
addition action only.